Nway弾などの追加

+++ Win32Project1/Win32Project1/Bullet2.h
@@ -15,6 +15,6 @@ public:
 	void MyDraw();
 	void MyPeculiarAction(BaseObject * obj);
 
-	int Speed;
+	double Speed;
 };
